requirements.txt
Django==5.2.2
finta==1.3
# gunicorn==23.0.0
gymnasium==1.1.1
numpy==2.2.5
pandas==2.2.3
plotly==6.1.2
pyarrow==20.0.0
PyMySQL==1.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
# sb3-contrib==2.6.0
scipy==1.15.3
stable-baselines3==2.6.0
yfinance==0.2.61
